Saving at this point.
There is still a fairly large AHRS bias when the blades are on that I haven't compensated for quite yet.

After discussion with William, the best approach is probably going to be to spike the corresponding values of the P matrix to the heading and heading bias, so that the filter will rely more heavily on the wheel odometry during the transient period right after the blades are turned on.